Introduce displayName placeholder for @ParameterizedTest name
Prior to this commit, it was not possible to include the display name
of a @ParameterizedTest method in the display name of individual
invocations of that parameterized test.

This commit introduces a '{displayName}' placeholder for the 'name'
attribute in @ParameterizedTest that allows developers to include the
display name of the @ParameterizedTest method in a custom display name
for invocations of that parameterized test.

This aligns with the existing '{displayName}' placeholder support for
@RepeatedTest.

Issue: #1496